Better to keep the types as they should be in the remote, fixes compilation of amuleGUI, why is the PartFile.h defining the tags for use on the remote gui? Doesn't seem to make much sense to me?

+++ src/PartFile.h
@@ -153,9 +153,9 @@ public:
 	size_t	GetSourceCount() const		{ return m_SrcList.size(); }
 	size_t	GetSrcA4AFCount() const		{ return m_A4AFsrclist.size(); }
 #else
-	size_t m_source_count, m_a4af_source_count;
-	size_t	GetSourceCount() const		{ return m_source_count; }
-	size_t	GetSrcA4AFCount() const		{ return m_a4af_source_count; }
+	uint16	m_source_count, m_a4af_source_count;
+	uint16	GetSourceCount() const		{ return m_source_count; }
+	uint16	GetSrcA4AFCount() const		{ return m_a4af_source_count; }
 #endif
 	uint16	GetTransferingSrcCount() const	{ return transferingsrc; }
 	uint16  GetNotCurrentSourcesCount()	const	{ return m_notCurrentSources; };
